separate SideServer into several menu classes
